Move all the extensions to AndroidNotificationExtension.
