+++ com.unity.mobile.notifications/Runtime/Android/AndroidNotificationExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b97d256371bb62248bb9f4404a72030a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
